// Front - TFG Hotel/src/app/core/interfaces/card-habitacion.interface.ts
export interface CardHabitacionInterface {
    titulo?: string;
    categoria?: string;
    descripcion?: string;
    imgSrc?: string;
    imgID?: number;
    enlace?: string;
    enlace_url?: string;

    tamaño?: number;
    imgHabitacionBase64?: string;
    precio?: number;
    cantidadDeHabitacionesDisponibles?: number;
}